+++ src/Method/Payment/SendInvoice.php
@@ -1,0 +1,109 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method\Payment;
+
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+use Vjik\TelegramBot\Api\Type\InlineKeyboardMarkup;
+use Vjik\TelegramBot\Api\Type\Message;
+use Vjik\TelegramBot\Api\Type\Payment\LabeledPrice;
+use Vjik\TelegramBot\Api\Type\ReplyParameters;
+
+/**
+ * @see https://core.telegram.org/bots/api#sendinvoice
+ */
+final readonly class SendInvoice implements TelegramRequestWithResultPreparingInterface
+{
+    /**
+     * @param LabeledPrice[] $prices
+     * @param int[] $suggestedTipAmounts
+     */
+    public function __construct(
+        private int|string $chatId,
+        private string $title,
+        private string $description,
+        private string $payload,
+        private string $currency,
+        private array $prices,
+        private ?int $messageThreadId = null,
+        private ?string $providerToken = null,
+        private ?int $maxTipAmount = null,
+        private ?array $suggestedTipAmounts = null,
+        private ?string $startParameter = null,
+        private ?string $providerData = null,
+        private ?string $photoUrl = null,
+        private ?int $photoSize = null,
+        private ?int $photoWidth = null,
+        private ?int $photoHeight = null,
+        private ?bool $needName = null,
+        private ?bool $needPhoneNumber = null,
+        private ?bool $needEmail = null,
+        private ?bool $needShippingAddress = null,
+        private ?bool $sendPhoneNumberToProvider = null,
+        private ?bool $sendEmailToProvider = null,
+        private ?bool $isFlexible = null,
+        private ?bool $disableNotification = null,
+        private ?bool $protectContent = null,
+        private ?string $messageEffectId = null,
+        private ?ReplyParameters $replyParameters = null,
+        private ?InlineKeyboardMarkup $replyMarkup = null,
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'sendInvoice';
+    }
+
+    public function getData(): array
+    {
+        return array_filter(
+            [
+                'chat_id' => $this->chatId,
+                'message_thread_id' => $this->messageThreadId,
+                'title' => $this->title,
+                'description' => $this->description,
+                'payload' => $this->payload,
+                'provider_token' => $this->providerToken,
+                'currency' => $this->currency,
+                'prices' => array_map(
+                    static fn(LabeledPrice $price) => $price->toRequestArray(),
+                    $this->prices,
+                ),
+                'max_tip_amount' => $this->maxTipAmount,
+                'suggested_tip_amounts' => $this->suggestedTipAmounts,
+                'start_parameter' => $this->startParameter,
+                'provider_data' => $this->providerData,
+                'photo_url' => $this->photoUrl,
+                'photo_size' => $this->photoSize,
+                'photo_width' => $this->photoWidth,
+                'photo_height' => $this->photoHeight,
+                'need_name' => $this->needName,
+                'need_phone_number' => $this->needPhoneNumber,
+                'need_email' => $this->needEmail,
+                'need_shipping_address' => $this->needShippingAddress,
+                'send_phone_number_to_provider' => $this->sendPhoneNumberToProvider,
+                'send_email_to_provider' => $this->sendEmailToProvider,
+                'is_flexible' => $this->isFlexible,
+                'disable_notification' => $this->disableNotification,
+                'protect_content' => $this->protectContent,
+                'message_effect_id' => $this->messageEffectId,
+                'reply_parameters' => $this->replyParameters?->toRequestArray(),
+                'reply_markup' => $this->replyMarkup?->toRequestArray(),
+            ],
+            static fn(mixed $value): bool => $value !== null,
+        );
+    }
+
+    public function prepareResult(mixed $result): Message
+    {
+        return Message::fromTelegramResult($result);
+    }
+}