+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - New #79: Add `answerInlineQuery` method, `InlineQueryResultsButton`, `InlineQueryResult*`, `InputMessageContent*` and
   `LabeledPrice` types.
 - New #80: Add `answerWebAppQuery` method and `SentWebAppMessage` type.
+- New #81: Add `sendInvoice` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.